Add full ability, Role and Permission resource
The first user(that will be the admin) will have all the access to all
resources by default, and add Role and Permission as manageable
resources.
These two new resources were added because the client of this product
won't be giving access to all users, he will probably hire one guy that
will handle all the access on the service, and that guy can give others
users access and leave the admin with any responsibilities besides taking
care of the guy that he gave permission.