Add logic to filter unauthorized users
* Why was this change necessary?
The need to filter unauthorized users to access or modify data so the
Admin can dynamically authorize a user.

* How does it address the problem?
By adding the function deny_access that redirects the user to the root(
or :back), and executing it inside a case
statement, and then it authorizes by the action that was called and
verying if the user has ability to access it.

* Are there any side effects?
Custom actions may be a future problem.