Fix performance and add authorization tests
* Why was this change necessary?
Better performance in lamps_controller_spec and the need to test
authorization in one of the controllers to make sure it's working.

* How does it address the problem?
By adding two class variables and setting them up in a before all block
so they ran once and it's accessible to every test.
Authorization tests were added in a case statement to minimize code but
they work just like the others tests in this file.

* Are there any side effects?
Custom actions are not supported in authorization tests and every action
will have a user signed in with a role and full permission unless you
called again in a test with nil.
The authorization tests have the expect be_success and be_redirect
because the URL may be different in the future.