Create firmware scaffold with green tests.
* Why was this change necessary?
The need to storage a binary file that can be served as the client
requests.

* How does it address the problem?
Creating a resource named firmware that delivers the file uploaded in
parts, requested by the client using the params 's' and 'e' in the show
action representing the range in the length of the file.

* Are there any side effects?
No.