Optimize lamp_stats table and rename serial_num column
* Why was this change necessary?
Improve performance and reducing useless space.
The serial_num was because rails don't use camelCase. I will add the
others fields in a separate branch

* How does it address the problem?
Adding index to serial_num and limiting some integer to smallint(except
the serial_num that needs to be bigint)

* Are there any side effects?
Better performance and new name to serial_num