Fix relationship between user and roles.
* Why was this change necessary?
Because the role doesn't belong to the user, the user that belongs to
the role.
When you create a new role you don't have to say to who that
role belongs, you just want to create a role that you can assign other
users.

* How does it address the problem?
By moving changing the foreign keys and the relationship in the models.
I also had to change the timestamps of the migrations so it can create
foreign keys from role and company to the user table
* Are there any side effects?
Not really, after I changed it I ran all the tests and got all green.