Add headers to firmwares show action and improve logic
* Why was this change necessary?
The headers are needed to reduce the payload of the response and offer
more info about the firmware.
The show action was a code refactoring.
Was need to redirect to the index page after save to prevent load the
content of the firmware.

* How does it address the problem?
Adding the method filter_header in the show action and adding new
headers with the checksum value and it's size

* Are there any side effects?
The refactoring code has no change.
All the headers added by rails in the controller were removed in the
show action.