Add new jbuilder response in networks#index
* Why was this change necessary?
Nodes was needed to render in the map with it's identifier, with the
link to the node and with it's the products and their links

* How does it address the problem?
The landing.js.coffee makes a call to networks#index.
The jbuilder will nest the objects in the json and add a translation to
the name of the final product so it's displayed nicely.

* Are there any side effects?
This generates a great amout of calls to a page( currently the root
page ) that will be accessed many times and it might be an overkill for
the database performance.