Fix Unpermitted name params and fix edit page reload
The user couldn't update his name because it wasn't configured in the
RegistrationController. Adding the before_filter with the method
configure_permitted_parameters did the trick.
Also it was needed a fix when the user inputs wrong or forgets a value
and the page reloads without layout. This was fixed by adding the line
41