Add Token Auth to Gateway in gateways table
Make gateway able to login and have a his session with token auth by
including devise_token_auth gem and some minor customizations.
The point is to make the user able to create a gateway with his
credentials and this action results in a functional endpoint for the
gateway make a POST to login in the system and have his token.