Remove uniqueness of uid + provider
It removes uniqueness of that index in gateways table in order to run
the migrations properly.
It will throw error of provider and uid being duplicated because uid is
blank in existing records.
So by removing it and runnign the migration, and then filling the uid
with a unique field in a second migration it will work just fine.